Move Workflow example to NSubstitute.

+++ examples/Workflow/WorkflowConsoleApp/Models.cs
@@ -1,11 +1,11 @@
 ﻿namespace WorkflowConsoleApp.Models 
 {
-    public record OrderPayload(string Name, double TotalCost, int Quantity = 1);
+    public record OrderPayload(string Name, decimal TotalCost, int Quantity = 1);
     public record InventoryRequest(string RequestId, string ItemName, int Quantity);
     public record InventoryResult(bool Success, InventoryItem orderPayload);
-    public record PaymentRequest(string RequestId, string ItemName, int Amount, double Currency);
+    public record PaymentRequest(string RequestId, string ItemName, int Amount, decimal Currency);
     public record OrderResult(bool Processed);
-    public record InventoryItem(string Name, double PerItemCost, int Quantity);
+    public record InventoryItem(string Name, decimal PerItemCost, int Quantity);
     public enum ApprovalResult
     {
         Unspecified = 0,
